[move-vm][closures] New reentrancy checker
This implements a new reentrancy checker that supports both the rules of existing native dispatch (fungible assets et. al) as the ones for function value dispatch. The rules for the lattter are described in [AIP-112](https://github.com/aptos-foundation/AIPs/blob/main/aips/aip-112.md#reentrancy-check), and are more liberal than for native dispatch: with funtion value dispatch, one can reenter a module, but resources defined in such a module are unaccessible. With native dispatch, reentrancy is already blocked when an attempt is made to enter a module the 2nd time.

The existing reentrancy check is pulled out from the interpreter into a new module `reentrancy_checker.rs` which handles both rule sets by processing when functions are entered/exited and whether by static call/dynamic dispatch.

There is no behavioral change, just a different implementation of a relative small logic (and now probably clearer than before), so no feature flag has been created for this. Existing tests for reentrancy are plenty and pass. The new logic is tested by a set of transactional tests.

Closes #15962